+++ 리아-이아현/4주차/mission02/src/pages/NowPlayingPage.jsx
@@ -1,12 +1,12 @@
 import MovieGrid from "../components/MovieGrid";
-import useCustomFetchMovies from "../hooks/useCustomFetchMovies";
+import useFetchMovies from "../hooks/useFetchMovies";
 
 const NowPlayingPage = () => {
   const {
     data: movies,
     isLoading,
     isError,
-  } = useCustomFetchMovies(`/movie/now_playing?language=ko-KR&page=1`);
+  } = useFetchMovies(`/movie/now_playing?language=ko-KR&page=1`);
 
   if (isLoading && movies.length === 0) {
     return (
